scripts/requirements.txt
Pillow
numpy
piexif